+++ storage/spider/hs_client/hstcpcli.cpp
@@ -3,7 +3,7 @@
 
 /*
  * Copyright (C) 2010-2011 DeNA Co.,Ltd.. All rights reserved.
- * Copyright (C) 2011 Kentoku SHIBA
+ * Copyright (C) 2011-2017 Kentoku SHIBA
  * See COPYRIGHT.txt for details.
  */
 
@@ -497,7 +497,7 @@ hstcpcli::response_recv(size_t& num_flds_r)
     char *const err_begin = start;
     read_token(start, finish);
     char *const err_end = start;
-    String e = String(err_begin, (uint32)(err_end - err_begin), &my_charset_bin);
+    String e = String(err_begin, err_end - err_begin, &my_charset_bin);
     if (!e.length()) {
       e = String("unknown_error", &my_charset_bin);
     }
